+++ src/main/java/in/koreatech/koin/domain/timetableV2/dto/TimetableFrameCreateRequest.java
@@ -1,9 +1,9 @@
-package in.koreatech.koin.domain.timetable.dto;
+package in.koreatech.koin.domain.timetableV2.dto;
 
 import static io.swagger.v3.oas.annotations.media.Schema.RequiredMode.REQUIRED;
 
 import in.koreatech.koin.domain.timetable.model.Semester;
-import in.koreatech.koin.domain.timetable.model.TimetableFrame;
+import in.koreatech.koin.domain.timetableV2.model.TimetableFrame;
 import in.koreatech.koin.domain.user.model.User;
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
