+++ src/main/java/in/koreatech/koin/domain/timetableV2/dto/TimetableLectureUpdateRequest.java
@@ -1,4 +1,4 @@
-package in.koreatech.koin.domain.timetable.dto;
+package in.koreatech.koin.domain.timetableV2.dto;
 
 import static io.swagger.v3.oas.annotations.media.Schema.RequiredMode.NOT_REQUIRED;
 import static io.swagger.v3.oas.annotations.media.Schema.RequiredMode.REQUIRED;
