+++ src/main/java/in/koreatech/koin/domain/timetableV2/model/TimetableFrame.java
@@ -1,13 +1,14 @@
-package in.koreatech.koin.domain.timetable.model;
+package in.koreatech.koin.domain.timetableV2.model;
 
+import static jakarta.persistence.CascadeType.ALL;
 import static jakarta.persistence.GenerationType.IDENTITY;
 import static lombok.AccessLevel.PROTECTED;
 
-import static jakarta.persistence.CascadeType.ALL;
 import java.util.List;
 
 import org.hibernate.annotations.Where;
 
+import in.koreatech.koin.domain.timetable.model.Semester;
 import in.koreatech.koin.domain.user.model.User;
 import in.koreatech.koin.global.domain.BaseEntity;
 import jakarta.persistence.Column;
@@ -15,9 +16,9 @@ import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.Id;
+import jakarta.persistence.Index;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
-import jakarta.persistence.Index;
 import jakarta.persistence.OneToMany;
 import jakarta.persistence.Table;
 import jakarta.validation.constraints.NotNull;
