+++ src/main/java/in/koreatech/koin/domain/timetableV2/repository/SemesterRepository.java
@@ -1,0 +1,23 @@
+package in.koreatech.koin.domain.timetableV2.repository;
+
+import java.util.List;
+import java.util.Optional;
+
+import org.springframework.data.repository.Repository;
+
+import in.koreatech.koin.domain.timetable.exception.SemesterNotFoundException;
+import in.koreatech.koin.domain.timetable.model.Semester;
+
+public interface SemesterRepository extends Repository<Semester, Integer> {
+
+    List<Semester> findAllByOrderBySemesterDesc();
+
+    Semester save(Semester semester);
+
+    Optional<Semester> findBySemester(String semester);
+
+    default Semester getBySemester(String semester) {
+        return findBySemester(semester)
+            .orElseThrow(() -> SemesterNotFoundException.withDetail("semester: " + semester));
+    }
+}
